move endpoing into gateway package
